+++ mlrun/__main__.py
@@ -27,12 +27,13 @@ import click
 import yaml
 from tabulate import tabulate
 
+import mlrun
+
 from .builder import upload_tarball
 from .config import config as mlconf
 from .db import get_run_db
 from .k8s_utils import K8sHelper
 from .model import RunTemplate
-from .platforms.iguazio import watch_stream
 from .projects import load_project
 from .run import get_object, import_function, import_function_to_dict, new_function
 from .runtimes import RemoteRuntime, RunError, RuntimeKinds, ServingRuntime
@@ -872,9 +873,11 @@ def clean(kind, object_id, api, label_selector, force, grace_period):
     is_flag=True,
     help="indicate the payload is json (will be deserialized)",
 )
-def stream_watcher(url, shard_ids, seek, interval, is_json):
+def watch_stream(url, shard_ids, seek, interval, is_json):
     """watch on a stream and print data every interval"""
-    watch_stream(url, shard_ids, seek, interval=interval, is_json=is_json)
+    mlrun.platforms.watch_stream(
+        url, shard_ids, seek, interval=interval, is_json=is_json
+    )
 
 
 @main.command(name="config")
